Adds a check before caching responses
This adds a check, _isSupportedUrl(), to the saveResponse() method of
DrupalLaravel Adapter so that only supported URLs (the Wordpress blog)
are cached with this particular caching mechanism. Previously, seemingly
all calls were being cached with this mechanism, which resulted in (for
instance) SSO calls being cached and therefore broken for all
authentication except for the account with tokens that are cached.