Product Category View Undefined Index Base
This error was a result of a missing handler (relationship) on the
product_groups View, which seemed to be some past image field. I removed
the relationship from the View and verified that the display of the View
remained the same. It did. I also verified that the error no longer was
logged. It was not.