[DEVELOPER-5646] Invalid Render Array Keys
This resolves an issue that prevented the following links (and
associated link text) from displaying for every Connector:
field_connector_link_1, field_connector_link_1_text,
field_connector_link_2, field_connector_link_2_text and
field_connector_details_url.

We were not properly accessing these values anyway, but in addition to
that, we were not verifying that any value(s) existed for any of the
fields before trying to access them, which resulted in both nothing
being rendered and many errors being thrown in the logs any time the
Connectors page was visited.

I am now verifying values exist, in a preprocessor for the
views-view-fields, and then accessingt those values and setting them in
the $variables array before accessing them in the Twig template.