URL Alias permissions changes
Content Creators can no longer edit or view URL Aliases, but I added the
ability for Content Admin to view and edit URL Aliases. NOTE: Content
Admin already had the ability to Administer URL Aliases, so this is a
somewhat redundant change in that sense. I thought it made sense to give
the permission to Content Admin though in case a quick edit had to be
made.