Install Trailing Slash and Fix All Products View
This installs the trailing slash (and dependencies: language) and sets
and configures some configuration. Also, configures the All Products
View such that the links to the Product overview pages have trailing
slashes. Apparently the Language module adds a config option to a ton of
config objects, hence the huge export.