Circumvent render arr, conditional trailing slash
This conditionally provides 2 ways that we will render the sidebar menu
Links on the product pages. One way for the legacy/export and a new way
for the MP build. These will be conditionally determined whether or not
the trailing_slash.settings.enabled config item is TRUE/FALSE.

The legacy/export process remains the same as it has been. The new
process though will, instead of providing a render array to the
template, just provide a title (string), url (string) and attributes
(Attribute object) to render each link so that we can manipulate the url
string to append a trailing slash.