Updates configuration checks for trailing_slash
I forgot that there are 2 'types' of config objects in Drupal 8, and
only 1 of those support config overrides. I was using the mutable config
object to check trailing_slash config key/values, but the mutable config
objects do not respect config overrides, which was discovered as an
issue once this went to MP. I have now changed the code so that I'm only
using the mutable object to inject URLs into the trailing_slash
settings, and instead using the immutable object to retrieve
trailing_slash config key/values.