[DEVELOPER-5679] Add all assemblies to new world
This adds all assemblies to Assembly fields on new world content types.
I have not included the Product Downlist List, Product Download Hero and
Product Try-It Hero assembly types to non-Product assembly fields only
because they will not work without an id to pass to download manager and
fetch the data necessary to render the assembly.