Fixed logic error with null-coalescing operator
This resolves an issue with a null-coalescing operator on 2 lines, which
both use the |capitlize filter, in
node--video-resource--learning-path-card.html.twig. An empty array was
being evaluated by the null-coalescing operator, so that was passed to
the capitalize filter, which results in an error since the capitalize
filter requires a string param.