[DEVELOPER-5758] Katacoda Breadcrumb Fix
To fix this issue with the missing intermediate breadcrumb link, we need
to add these pages, and their desired structure/nesting, to a menu. I
have created a new menu in Drupal that we can add this Katacoda content
to. When courses and individual scenarios are properly added to the
menu, this breadcrumb issue will be resolved.