DEVELOPER-5820 Dynamic Content List Assembly
This commit updates the Dynamic Content List Assembly to allow pulling
all latest content, ordered by most recent, from Drupal and Wordpress.
Currently, an editor has to specify some categories in order to get
results from Wordpress or Drupal. This makes it so that when no
categories are specified, the latest content is pulled from Drupal and
Wordpress.