Add Redirects and Katacoda types to Trailing Slash
This adds the Redirects, Katacoda Course and Katacoda Individual Lesson
entity types to the Trailing Slash configuration. The Katacoda content
types should have been added upon creation of the content types, so it
is good we are taking care of that now. Adding the redirects should mean
that redirect paths are now properly sent with the trailing slash.