DEVELOPER-5768: Add URL Alias for Content Creator
In DEVELOPER-5664, I removed the permission for Content Creators to
add/edit URL aliases. This was part of an effort to reduce clutter on
Node edit forms. At the time, I thought that the URL aliases for all
nodes were being generated automatically by Pathauto. That is not the
case, so this was a regression.

Furthermore, now that we are using Compose and the URL alias field is
not on the main page causing visual bloat, I do not think that this is
adding to UX clutter.