Update Katacoda Courses View and configuration
This reworks the way that the /courses page works, behind the scenes.
Previously, everything was 'hardcoded' because a Views page display was
being used to render the content. We will now use a Views block display
that will be configured to only display on the /courses route. This
allows us the flexibility to use a Page node to create our /courses
route so our editors can edit the content of the hero (Rich Text with
Image) assembly.

I have also added the Katacodas assembly type to all assembly reference
fields per Joe's request.