Still Loads Course Cards with 0 Lessons
This commit ensures that Katacoda Course Cards still load, with all the
data available, even if there are zero courses attached to the course. I
discovered this in the PR environment since there are multiple courses
with zero lessons.

This also resolves a CSS specificity issue that was causing an
unexpected issue on the Katacoda Course nodes (default view mode)
because I wasn't specifically targeting the Card view mode with my
styles.