Assembly Lessons using entity display
I discovered that the Katacoda Individual Lesson cards on the Katacoda
Course node/page (default view mode) was not actually rendering the
entity, but instead it was doing a custom rendering of them using
precessors, template, etc. We want to actually render the Katacoda
Individual Lesson (Card display mode) entities, so I've done that with
this commit.