DEVELOPER-5840: PSA-2019-05-07
This implements the Drupal core update from 8.6.15 to 8.7.1, which
includes a necessary Lighting update to 3.3.1. After making these
updates with a `composer update drupal/core acquia/lightning
--with-dependencies`, I rebuilt the cache and then ran `drush updb`,
then I rebuilt the cache and ran `drush update:lightning`. Lastly, I
rebuilt the cache, verified the site was loading and that there weren't
errors in the logs, and after verifying that, I exported config.