BUG: Uninitialized String Offset Product URL Alias
Within a validation hook, used to validate that non-Product URL aliases
will not conflict with programmatically-generated Product URL paths, we
are not validating that a URL alias is set prior to accessing the value.
This commit validates that the URL alias is set prior to accessing the
value.

This will remove the following error from the Drupal logs: "Notice:
Uninitialized string offset: 0 in..."