[MWES-2920] Ensure that search tests explictly wait for the results to be shown before making assertions
We're seeing intermittent failures on the search tests. Running through the tests by hand everything seems to be fine. Looking over the tests there doesn't seem to be anything that explicitly waits for the results to be loaded before running assertions. In the new preview environments, pages will load noticeably quicker than before as we have gzip compression enabled. It could simply be that the improved performance here is showing up race conditions that we've always had but were simply hidden by the performance of the site