[MWES-3311] Added context parameter to build job for managed platform CI pipeline
We can now send a context parameter to the CI pipeline on jenkins.paas.redhat.com. In our case this context will be either 'rebuild this please' or 'retest this please'.

When the context is 'rebuild this please', the CI pipeline will completely rebuild the preview environment associated with a pull request. When the context is 'retest this please' only the acceptance tests will be executed meaning that you can retry flaky tests much more efficiently. The default is for the context to be 'rebuild this please'.