[MWES-3338] Update Drupal to use /cache/drupal for storing the Twig Cache
In trying to understand our performance problems on Managed Platform, I was looking at where we are storing our Twig Cache and it turns out we're actually storing it on an NFS mount which is going to be horrendously slow. This update instructs Drupal to store the cache at /cache/drupal on the container filesystem. Each environment can then make a decision about how it wants to provide the /cache/drupal implementation and I suspect the best option for us in production and other mpaas environments is to provide a tempfs mount at /cache/drupal.

One caveat to this change is that it does make the assumption that Twig templates do not change in your multi-web-head environment. For us I believe this holds as our deployments are immutable, but it does mean that in the future we have to be aware of anything that would allow editors to make alterations to Twig templates on the fly