[MWES-3342] Migrate Drupal to use the database for configuration storage
This commit updates our bootstrap of Drupal to perform a one-time migration of our configuration from the filesystem to the database. We are doing this because currently we store our configuration on an NFS disk which is likely to be slow and the recommended approach is to store configuration in the database.

This commit uses Ansible to perform the migration in the safest way possible, ensuring that we have backups of everything as we do the migration