Create a getContentCompositeMultiple method
This creates a getContentCompositeMultiple method in the WordpressApi
service that will be used/called from the getContentByCategory method
since multiple results will be passed. This is better than the previous
implementation because in getContentCompositeMultiple, we will only make
1 network call to the WP REST API, whereas in the previous/existing
implementation there are N calls made where N is the number of WP posts
retrieved.

I also changed the apiUrl attribute from origin to d.r.c.