Product page drafts available to anonymous role
This resolves an issue where Product page drafts were available to view
by anonymous users (unauthenticated). Since we have a custom
implementation of a Node-like page, controller and routing, we need to
validate each possible scenario where a node is or isn't published and
the role(s) of the current user visiting the page. In this instance, we
we properly validating the roles, but did not check the published status
of the Node object. I thought that this was handled by the routing, if
the 'thing' you are routing is a Node object, but it does not.

In addition to these changes, there are some minor changes made to
resolve some variables that were being accessed without validation in
rhd_common.product_related.inc and rhdp.theme. This should resolve some
related errors in the Drupal watchdog logs.