Adds condition for paragraphs-based downloads page
This adds a conditional that will, when appropriate, serve the old,
paragraphs-based Product downloads page from the
ProductsDownloadsController. It feels like we should attempt to better
clarify the code that is executed/shared between the old,
paragraphs-based Product pages and the new, Assembly-based Product
pages. Currently, this implementation is very confusing when compared to
standard Drupal functionality due to the 2 Product pag implementations
and the paradigm-breaking Node behavior. Hopefully this can be
accomplished in tandem with some work to improve the performance of the
Product pages.