Display paragraphs-based Download for admin users
This now also serves the paragraphs-based product download page for
authenticated users with permissions to view all revisions. I also added
in a change to properly serve a 404 when no Getting Started content
exists for an old, paragraphs-based Product page. Previously, a page was
served with just an empty main content area.