Update drupal/bootstrap and set no CDN
This updates the drupal/bootstrap contrib theme to its latest version,
which is a significant improvement/change. This commit also sets the
config of the theme to declare no CDN, as opposed to the JSDelivr CDN.
Both of these changes are necessary to remove an unpredictably large
number of CDN requests that were being made on every admin request to
the JSDelivr CDN. As noted in the Github issue, this should result in a
significant performance increase in pages that use the admin theme(s).