[MWES-3419] Ensure that we map the current UID into the containers to avoid filesystem permission issues
This commit updates the local development experience to use the UID of the current user as the UID of the process that runs within the Docker containers within the local development environment. This should hopefully rectify any issues we have with the previous approach that used root to bootstrap the environment. In using root, we typically end up with files that are no longer owned by the current user and therefore are a pain to delete without using privilege escalation such as sudo