Download Component Libraries module, add Card
This downloads the Components Libraries module and adds a Card
component. It also implements an Assembly Static Item template in the
RHDP2 templates/ directory that will include the Card component. This is
just a proof of concept currently to demonstrate how we can create
reusable components across assembly types (or node types).