Replace rhd_guzzle_throttle with Drupal caching
This uninstalls the custom rhd_guzzle_throttle module and replaces the
caching that module provided, for external calls to our Wordpress API
and Download Manager, with a standard Drupal Cache API implementation.
This also includes a database update in rhd_common.update.inc to drop
the laravel_cache table as we no longer will use that and do not
otherwise need it.