[3012] Updated the local build scripts to temporarily request the font awesome licence
As a temporary work-around this commit updates the local build scripts to request the user to provide the font awesome licence key. This in turn is used within the Dockerfile.mp to build the new rhd-frontend implementation dist assets for the rhdp2 Drupal theme.

I say this is temporary because ultimately we want the rhd-frontend assets on a CDN. Not only does it add about 200Mb of bloat to our Drupal container, it should be handled by a completely separate build process so that Drupal can simply reference the files it needs at a known version from our CDN