Resolved build failures for rhdp2 theme
This resolves the errors that were preventing the Docker build from
successfully building dist assets the rhdp2 theme. We cannot sync the
themes/custom directory in the way we were because it will overwrite the
dist assets.

I have also made the rhdp2 the default theme in system.theme.yml and I
have pulled the latest version of rhd-frontend's master branch.