Delete rhd-frontend after building assets
I added some lines to our Dockerfile.mp to move the dist assets (CSS and
JS file(s)) we need into the css/ or js/ directories in the root of the
rhdp2 theme directory. After moving these dist assets, I added a line to
entirely delete the rhd-frontend directory/project from the filesystem.
As we only need the dist assets to reference, I do not see a reason to
keep unnecessary bloat and hopefully this will reduce the image size.