Fixes broken favicon paths & updates rhd-frontend
This fixes some broken favicon paths by moving the rhd-frontend/favicons
directory into the root of the rhdp2 theme prior to deleting the
rhd-frontend directory, and then we have updated the file references in
html.html.twig to meet this new directory structure. This commit also
updates the rhd-frontend package to its latest commit on the master
branch. The changes include a commit that improves the README, a commit
for Main Navigation reference implementation and a commit to remove the
multiple font references.