SURGE-3005: Rich Text assembly
This makes the Rich Text assembly type template and preprocessors work
such that the DX Surge design requirements are addressed. I had to move
some code from rhdp.theme to rhdp2.theme in the assembly preprocessor
for background images and audience types. These blocks of code apply to
assemblies generally, as opposed to specific assembly types, so I think
that it is alright to have this code in the preprocess for now, even
though we discussed  moving assembly-type-specific code into
AssemblyBuild plugins when possible.