SURGE-3056: Curated Links assembly type
Creates the Curated Links template in the rhdp2 theme according to the
UX designs and FE reference implementation. I think that there was some
confusion about exactly what the Curated Links assembly type (or perhaps
component) is, or perhaps just the difference between the component and
the assembly type is confusing. In Drupal, the assembly type is a banner
that consists of a list of links _and_ (it is not required, but in
almost all cases) additional content which can be the following assembly
types: Curated Content, Dynamic Content, Rich Text and Rich Text
Columns.

Since I needed the right circle arrow (FontAwesome) for this PR, I also
decided to include the new RHD FontAwesome kit in rhdp2.libraries.yml.
This should resolve #3070.