[MWES-3422] Add static JSON from legacy Docroot
Adding this piece of JSON from the legacy docroot that we currently mount from the static web servers. No-one seems to know what it's for, but until we can track that down it will prevent us from decomissioning the legacy web servers. Instead we can simply host it as part of Drupal (for now) and that will allow us to decommission the old infrastructure.