Card Grid Base Implementation
This implementation a Card Grid component in the rhdp2/components
directory that other templates with include. I have included a few
examples of this in this commit since I had to implement them anyway to
test this work. This does not represent 'feature complete' state of any
of these components, but it provides an example of how we can integrate
the card grid component that was merged into the FE repo this morning.