Revise author rendering strategy in Card component
This allows authors to be rendered in the Article card by revising the
way that we render, and in the instance of author entity references
preprocess, the author content. Using this new approach, we can support
both instances I see currently in our content model: 1) Including
authors on a piece of content via a plain text field and 2) Including
authors on a piece of content via an entity reference to the Authors
bundle (see Articles).