Update rhd-frontend and fix CTAs in Card Grid
This updates the rhd-frontend package to the latest commit on `master`,
and it also makes the Card Grid component use the latest reference
implementation (pushed yesterday by Adela). As a follow up, I have
created a ticket in the `rhd-frontend` repo for UX to define how they
want multiple CTAs in a Card Grid to be formatted.