[MWES-3564] Capture acceptance tests job in a Jenkins pipeline
This commit is the first part of the work to be able to run the acceptance tests as a Jenkins pipeline. In order to do that I've had to tidy things up a little so that we don't have to run a bunch of Ruby commands. In addition, the README.md did not reflect real-life and the tests did not run following the instructions in there.

So the major changes are:

All of the test commands are now explicilty defined in package.json which means that they can be easily executed locally simply by running npm commands.
Created a docker-compose definition for running the said npm commands in a container and this is what will be used on Jenkins as part of the pipeline

There is still a bit of work to make this backwards compatible so that the existing testing infrastructure continues to work as is.