Remove display_submitted section
This removes a conditional section of the template that would never be
rendered since we have the display_submitted unchecked on the topic_page
node form, and presumably will always want this to be unchecked given
the purpose  of the topic pages. They are not content that is really
tied  to an author in a meaninful way like blog posts can be.