Updates Card Grid CTA Markup
This updates the Card Grid CTA markup to match the latest changes to the
rhd-frontend reference implementation. See: [SURGE-238] Multiple CTAs in
Card Grid Component.

With this change, there are 32px of vertical space between the last row
of cards and the CTAs, and there are 24px of horizontal space between
each CTA.

You can verify this change at the bottom of this page:
/harmonize/styleguide/