Preserve existing visual styles
Matt Moen pointed out an issue in my PR where existing visual styles
would be removed during the update hook. This resolves that issue by
getting the visual styles as an array, properly removing the 1 class we
are targeting, adding the 2 classes and setting the value  of the
visual_styles field with that modified array.