Includes rhd.old.min.js and Angular and settings
This includes the rhd.old.min.js file, Angular dependencies (required
for rhd.old.min.js) and the necessary Drupal settings for
rhd.old.min.js. rhd.old.min.js provides a lot of functionality, but the
piece relevant to this particular PR is the 'More Like This' section.
Currently, this section will render with Foundation classes i.e. the
grid will be wrong and the Card styles do not match our current Card
component on the site. However, I created a ticket in the FE repo here:
https://github.com/redhat-developer/rhd-frontend/issues/305 to update
the Foundation classes with PatternFly classes.