Load rhd.min.js in rhdp2 theme
This adds the building of rhd.min.js file to the FE build process in
Dockerfile.mp and dev/build-theme.sh and the copying of the file into
the ./rhdp2/js directory. It also declares the file in the
rhdp2.libraries.yml file. Lastly, I've added rhd.min.js to .gitignore
since we do not want to track a file we are building.