SURGE-3047: Wordpress Post Reference
This adds the proper template suggestion(s) and templates to render the
Wordpress Post Reference assembly type as a Card. The actual wordpress
post will render as a Tile, which you will see configured and in the
Twig template name. I am doing this, instead of using the Card because
this implementation has to render the _content_ of the Card vs. the
entire Card, so the markup is actually different. The top-level
container element comes from assembly--wordpress-post-reference.