SURGE-3178: Coding Resources
This implements the Coding Resources node type, full/default view mode.
This includes the Twig template and CSS changes. I also had to remove
the contextual loading of the HighlightJS library, and instead load it
everywhere, because the non-Assembly-based <pre><code> blocks now depend
on HighlightJS for styling and highlighting.