Display Code Snippet Assemblies
A regression was accidentally introduced here:
https://github.com/redhat-developer/developers.redhat.com/commit/7749cc4011c411c5f31f2c46b55b4e5436101d53#diff-304ef1e9f8b0e28569a944480f0512da
that removed a necessary variable for rendering the code snippet
assembly. I have added back the preprocessing of that variable, while
still leaving out the attachment of the highlight-js library, as that is
included globally to accommodate all <code><pre> elements. I used UX's
Marvel design to determine the 8px of padding.