[#3304] Remove multiple builds of rhd-frontend theme
This commit does a couple of things.

Firstly it removes any reference to the fontawesome stuff from the Drupal build. We are now hosting those assets from a CDN.

Secondly it only calls npm run-script build to create the new rhd-frontend assets. We were previously calling npm run-script styles and then npm run-script build, however the build script additionally calls styles, so this was a duplication