[MWES-3660] Migrate to merged SSO servers
This is the first step in migrating to the new SSO servers hosted by IT. In this commit I'm simply exposing the realm and client_id as additional configuration parameters that will facilitate an easier migration because things no longer need to be hard coded. This commit does not alter the current approach to login or the SSO servers in use.