[MWES-3660] Migrate to new SSO servers
Phase 2 of the migration work. This updates local dev, dev and stage environments to use the new SSO servers. This will give us an opportunity to test that the new SSO flows work as expected before migrating production.

Additionally this updates the front-end JS to make use of the configuration variables exposed in part 1 of this migration