Remove the rhd-frontend git submodule, keep files
This removes the rhd-frontend git submodule while keeping the
rhd-frontend files in the exact same place. This commit is a result of
many conversations over the past few weeks about how we can improve our
FE dev experience and generally the dev experience and QA experience for
everyone who contributes to this project. This commit can be considered
a Phase 0 step towards that goal. We have documented the requirements in
DX Reboot: Defining a desirable FE dev experience in Google Docs.