[MWES-3562] Add Jenkinsfile to support migration of PR build trigger to PSI Jenkins
This commit adds a new Jenkinsfile to the project root to support our migration to PSI Jenkins instance for triggering of pull request builds. It does not
replace the current Jenkinsfile on the project as the migration needs to be phased and we also need the ability to rollback if things don't work out.

A further update will merge these changes into the main Jenkinsfile and remove the Jenkinsfile.psi once the migration to our PSI Jenkins instance is complete
and verified.