DEVELOPER-3394: Katacoda course header spacing
This commit ensures that the spacing of the hero component on
`/courses/[course-slug]` matches the spacing of the hero component on
`/courses`. While working on this, I have also included some clean up of
the Katacoda implementation. This clean up includes:

- the removal of a lot of unused preprocessing code in
rhd_katacoda_preprocess_node(),
- fixing the function declaration and logic in
rhd_katacoda_preprocess_breadcrumb() (previously
rhdp_preprocess_breadcrumb which would never execute in this context),
- the preprocessing of the subtitle on Katacoda Course cards is now handled in
the preprocessor instead of in the template.