[No Issue] Integrate running the WordPress migrations into Drupal boot sequence
This commit updates the Drupal boot sequence to optionally run our WordPress migrations if the WORDPRESS_RUN_MIGRATIONS feature flag is set to true. If this is the case then the WordPress migration code will be executed.

Additionally rhd.settings.php is updated to include a WordPress database definition so that the migrations can pull data from the source to import into Drupal.