Display Image field, hide Media reference
This config export will now make it so that the Paragraph with Image
paragraph type displays its Image field on both the paragraph edit form
and when viewing the paragraph. It also removes the Media reference
field from display on the form and when viewing the Paragraph. There is
a 0-day bug with the implementation of the Media reference field, from
what I can tell dating back to June 13th, 2017, that makes the field
widget display on the paragraph edit form but entirely fails to render
it when viewing the paragraph due to our completely custom rendering
process of the old, Paragraphs-based Product pages.