Set required cron interval ImageCache External
There is a required cron interval config keyvalue that we apparently had
never set for the ImageCache External module. I tested this against the
1.1 and 1.2 releases locally and it _seemed_ like the field was required
in both releases. I am not sure, at least at this point in my debugging,
why we started to encounter these cron failures at this point, instead
of at an earlier date whenever this required value was introduced to the
contrib module. Setting a value resolves the error that arises during
cron execution though. I will verify this on the PR build once a build
either fails or completes successfully by reviewing the logs and testing
the command(s) against the commandline of the PR environment.