Fix theme suggestions conditional logic
The conditional logic we have in hook_theme_suggestions_alter() is not
properly validating that we have retrieved a Node object in $node. I
have adjusted the logic to ensure that we've actually fetched an object
that is an instance of NodeInterface. This should resolve an issue that
arises when trying to view Node revisions.