Drupal core patch to properly execute update hooks
This includes a Drupal core patch "system.schema can end up with missing
schema information for some modules, resulting in hook_update_N() not
getting called" that wil result in hook_update_N() being executed
properly/as intended for a number of modules that, via this bug, did not
have entries in the key-value table as system.schema, as designed in
Drupal core.

As a consequence, i have updated a few files that contain update hooks,
and I've exported on linkit config file that was generated as a result
of the update hook(s).