Fix empty cid for Wordpress API without category
There was a scenario where a dynamic assembly, fetching content from the
Wordpress API, does not have to specify a category. In this scenario, if
content is fetched with the method getContentByCategory(), it would
result in an empty string being used as the cid. Instead, I have set it
to a string like `wordpress_api:by_category:per_page_4-logic_or` which
is the same as other cid generate in the method, but with the category
ommitted. With this new, fixed cid, we can cache this data for calls
with the same values for 'per page' and 'logic'.