Reduce TTL of Download Manager and Wordpress API
I am going to reduce the TTL of our configurable Download Manager and
Wordpress API ttl config objects because of how painful it currently is
to rebuild Drupal's cache on Prod, due to the rhd_guzzle_throttle issue.
Even though this PR should resolve that error, I would rather start with
this lower TTL and bump it up if everything performs as we expect on the
PR environment.