[RHDX-448] Uninstall jsonapi and schemeta
We are no longer using the functionality of these modules, so we should
remove them. Removing these modules should remove a ton of routes from
the router table. After these routes are removed, we should continue to
monitor the recurring issues we've experienced that note DB queries to
the router table