RHDX-471: Related Products field renders as Card
The related products field on the Books node type now renders the
referenced Product nodes with the Card view mode. This required some
minor changes to make the Card view mode, for the Product node type,
display as intended.

I also decided to have the Featured Products assembly type render the
referenced Products using the Card view mode since it is more intuitive
that the Featured Tile view mode, and we eventually want to eliminate
card-related usage of the Featured Tile or Tile view modes in favor of
Card.