RHDX-465 Compact Dynamic Article List Length
This adds a field to the Compact Dynamic Article List assembly type, the
same field that is on the Dynamic Content Feed, that allows an editor to
configure the number of articles displayed in the list. The current
number of articles displayed is 8. This provides a default of 8, but
allows an editor to set the valuable anywhere between 4 to 10.